Fix fluentd elasticsearch gems
The elasticsearch gems that come bundled with td-agent don't work with
the OSS elasticsearch version any longer, so we need to downgrade them.

Patches are under review upstream. Cherry-pick for stable branches only
does the gem downgrade, for some reason yet unknown to me, td-agent is
unpinned in our builds already but not upstream.

[0] https://review.opendev.org/c/openstack/kolla/+/823071/7
[1] https://review.opendev.org/c/openstack/kolla/+/821695/11

Signed-off-by: Dr. Jens Harbott <harbott@osism.tech>