Get spice-html5 from git (#453)
The last release of spice-html5 was some time ago and in Ubuntu 22.04
there is only the 0.2.2 package available. A fix for an issue with
keyboard mappings for international keyboard layouts was not yet release
at all. To be able to solve this issue we now get the spice-html5 files
from the Git repository.

https://gitlab.freedesktop.org/spice/spice-html5/-/merge_requests/14

Part of osism/issues#1039

Signed-off-by: Christian Berendt <berendt@osism.tech>