2024.2/2025.1: add backport-964771.patch
Set upper_bound column in bios_settings table to bigint

Currently the column is set to a signed int, causing issues
for systems reporting larger values. The value of upper_bound
can be up to 32-bit unsigned as per redfish standard.

This patch will change the data type to BigInteger, which will
allow the larger values to be stored.

https://review.opendev.org/c/openstack/ironic/+/964771

Signed-off-by: Christian Berendt <berendt@osism.tech>